+++ pom.xml
@@ -43,8 +43,8 @@
         <module>enum-spring-boot-starter</module>
     </modules>
     <properties>
-        <revision>0.9.0</revision>
-        <admin4j-dependencies.version>0.9.2</admin4j-dependencies.version>
+        <revision>0.9.3-SNAPSHOT</revision>
+        <admin4j-dependencies.version>0.9.3-SNAPSHOT</admin4j-dependencies.version>
         <maven.compiler.source>8</maven.compiler.source>
         <maven.compiler.target>8</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
