+++ web-spring-boot-starter/pom.xml
@@ -16,11 +16,6 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.alibaba</groupId>
-            <artifactId>transmittable-thread-local</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
             <groupId>com.admin4j.common</groupId>
             <artifactId>admin4j-common</artifactId>
         </dependency>
