+++ web-spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   com.admin4j.framework.web.autoconfigure.JacksonAutoConfiguration,\
-  com.admin4j.framework.web.autoconfigure.WebAutoConfiguration,\
-  com.admin4j.framework.web.autoconfigure.UserContextAutoConfiguration
+  com.admin4j.framework.web.autoconfigure.WebAutoConfiguration
